requirements.txt
aiohttp-requests
google-cloud-speech

